LPS-116569 Split into files, one for the logic and two for the UI
