+++ modules/apps/ratings/ratings-taglib/src/main/resources/META-INF/resources/js/components/RatingsStars.js
@@ -12,17 +12,14 @@
  * details.
  */
 
-import ClayButton from '@clayui/button';
-import ClayDropDown from '@clayui/drop-down';
-import ClayIcon from '@clayui/icon';
-import ClayLayout from '@clayui/layout';
-import classNames from 'classnames';
 import {useIsMounted} from 'frontend-js-react-web';
 import PropTypes from 'prop-types';
-import React, {Fragment, useCallback, useState} from 'react';
+import {useCallback, useState} from 'react';
 
 import {TYPES} from '../constants';
 import Lang from '../utils/lang';
+import RatingsSelectStars from './RatingsSelectStars';
+import RatingsStackedStars from './RatingsStackedStars';
 
 const SCORE_UNVOTE = -1;
 
@@ -138,29 +135,25 @@ const RatingsStars = ({
 		return Lang.sub(srAverageMessage, [averageScore, numberOfStars]);
 	};
 
-	return type !== TYPES.STACKED_STARS
-		? RatingsSelectStars({
-				averageScore,
-				disabled,
-				getSrAverageMessage,
-				getTitle,
-				numberOfStars,
-				onVote: handleVote,
-				score,
-				starScores,
-				totalEntries,
-		  })
-		: RatingsStackedStars({
-				averageScore,
-				disabled,
-				getSrAverageMessage,
-				numberOfStars,
-				onVote: handleVote,
-				randomNamespace,
-				score,
-				starScores,
-				totalEntries,
-		  });
+	const RatingsStarsTypes = {
+		[TYPES.STACKED_STARS]: RatingsStackedStars,
+		[TYPES.STARS]: RatingsSelectStars,
+	};
+
+	const RatingsStarsUI = RatingsStarsTypes[type];
+
+	return RatingsStarsUI({
+		averageScore,
+		disabled,
+		getSrAverageMessage,
+		getTitle,
+		numberOfStars,
+		onVote: handleVote,
+		randomNamespace,
+		score,
+		starScores,
+		totalEntries,
+	});
 };
 
 RatingsStars.propTypes = {
@@ -176,240 +169,3 @@ RatingsStars.propTypes = {
 };
 
 export default RatingsStars;
-
-function RatingsStackedStars({
-	averageScore,
-	disabled,
-	getSrAverageMessage,
-	numberOfStars,
-	onVote,
-	randomNamespace,
-	score,
-	starScores,
-	totalEntries,
-}) {
-	return (
-		<fieldset className="rating ratings-stacked-stars ratings-stars">
-			<div
-				className="ratings-stars-average"
-				title={averageScore.toFixed(1)}
-			>
-				<span className="inline-item inline-item-before">
-					{starScores.map(({label: score}, index, arr) => {
-						const previousScore =
-							arr[index - 1] && arr[index - 1].label;
-						let symbol = 'star-o';
-
-						if (averageScore >= score) {
-							symbol = 'star';
-						}
-						else if (averageScore > previousScore) {
-							symbol = 'star-half';
-						}
-
-						return (
-							<ClayIcon
-								className="ratings-stars-average-icon"
-								key={score}
-								symbol={symbol}
-							/>
-						);
-					})}
-				</span>
-				<span className="inline-item ratings-stars-average-text">
-					{!!totalEntries &&
-						` (${totalEntries} ${
-							totalEntries === 1
-								? Liferay.Language.get('vote')
-								: Liferay.Language.get('votes')
-						})`}
-				</span>
-				<span className="sr-only">{getSrAverageMessage()}</span>
-			</div>
-
-			<div
-				className={classNames({disabled}, 'ratings-stacked-stars-vote')}
-			>
-				<div className="ratings-stacked-stars-vote-stars">
-					{starScores.reverse().map(({label, value}, index) => {
-						const id = `${randomNamespace}star${label}`;
-						const srMessage =
-							index === 0
-								? Liferay.Language.get(
-										'rate-this-x-star-out-of-x'
-								  )
-								: Liferay.Language.get(
-										'rate-this-x-stars-out-of-x'
-								  );
-						const full = label <= score;
-
-						return (
-							<Fragment key={value}>
-								<input
-									checked={label === score}
-									className="sr-only"
-									disabled={disabled}
-									id={id}
-									name={`${randomNamespace}rating`}
-									onChange={() => {
-										onVote(index);
-									}}
-									type="radio"
-									value={value}
-								/>
-								<label
-									className={
-										full ? 'ratings-stars-star-full' : ''
-									}
-									htmlFor={id}
-								>
-									<ClayIcon
-										className="ratings-stars-icon-full"
-										symbol={'star'}
-									/>
-									<ClayIcon
-										className="ratings-stars-icon-empty"
-										symbol={'star-o'}
-									/>
-									<span className="sr-only">
-										{Lang.sub(srMessage, [
-											label,
-											numberOfStars,
-										])}
-									</span>
-								</label>
-							</Fragment>
-						);
-					})}
-				</div>
-
-				{score !== 0 && (
-					<ClayButton
-						className="lfr-portal-tooltip ratings-stacked-stars-delete"
-						disabled={disabled}
-						displayType="unstyled"
-						onClick={onVote}
-						title={Liferay.Language.get('delete')}
-					>
-						<ClayIcon symbol="times-circle" />
-					</ClayButton>
-				)}
-			</div>
-		</fieldset>
-	);
-}
-
-function RatingsSelectStars({
-	averageScore,
-	disabled,
-	getSrAverageMessage,
-	getTitle,
-	numberOfStars,
-	onVote,
-	score,
-	starScores,
-	totalEntries,
-}) {
-	const [isDropdownOpen, setIsDropdownOpen] = useState(false);
-
-	const handleOnClick = (index) => {
-		setIsDropdownOpen(false);
-		onVote(index);
-	};
-
-	return (
-		<ClayLayout.ContentRow
-			className="ratings ratings-stars"
-			noGutters
-			verticalAlign="center"
-		>
-			<ClayLayout.ContentCol>
-				<ClayDropDown
-					active={isDropdownOpen}
-					menuElementAttrs={{
-						className: 'ratings-stars-dropdown',
-					}}
-					onActiveChange={(isActive) => setIsDropdownOpen(isActive)}
-					trigger={
-						<ClayButton
-							aria-pressed={!!score}
-							borderless
-							className="ratings-stars-dropdown-toggle"
-							disabled={disabled}
-							displayType="secondary"
-							small
-							title={getTitle()}
-							value={score}
-						>
-							<span className="inline-item inline-item-before">
-								<ClayIcon symbol={score ? 'star' : 'star-o'} />
-							</span>
-							<span className="inline-item ratings-stars-button-text">
-								{score || '-'}
-							</span>
-						</ClayButton>
-					}
-				>
-					<ClayDropDown.ItemList>
-						{starScores.map(({label}, index) => {
-							const srMessage =
-								index === 0
-									? Liferay.Language.get(
-											'rate-this-x-star-out-of-x'
-									  )
-									: Liferay.Language.get(
-											'rate-this-x-stars-out-of-x'
-									  );
-
-							return (
-								<ClayDropDown.Item
-									active={label === score}
-									key={index}
-									onClick={() => {
-										handleOnClick(index);
-									}}
-								>
-									{label}
-									<span className="sr-only">
-										{Lang.sub(srMessage, [
-											index + 1,
-											numberOfStars,
-										])}
-									</span>
-								</ClayDropDown.Item>
-							);
-						})}
-
-						<ClayDropDown.Item
-							disabled={score === 0}
-							onClick={handleOnClick}
-						>
-							{Liferay.Language.get('delete')}
-						</ClayDropDown.Item>
-					</ClayDropDown.ItemList>
-				</ClayDropDown>
-			</ClayLayout.ContentCol>
-
-			<ClayLayout.ContentCol>
-				<span className="ratings-stars-average">
-					<span className="inline-item inline-item-before">
-						<ClayIcon
-							className="ratings-stars-average-icon"
-							symbol="star"
-						/>
-					</span>
-					<span className="inline-item ratings-stars-average-text">
-						{averageScore.toFixed(1)}
-						{!!totalEntries &&
-							` (${totalEntries} ${
-								totalEntries === 1
-									? Liferay.Language.get('vote')
-									: Liferay.Language.get('votes')
-							})`}
-					</span>
-					<span className="sr-only">{getSrAverageMessage()}</span>
-				</span>
-			</ClayLayout.ContentCol>
-		</ClayLayout.ContentRow>
-	);
-}
